Consider go type name when autobinding
Currently, generated schema type names are normalized, for instance - SomeTYPE in the schema will be generated as SomeType in the model.

When autobinding, however, we only consider the schema type name when searching for it in the relevant package(s),
thus type names that differ post normalizations aren't auto-bound properly and are instead re-generated.

This commit suggests a fix where we'd try to autobind for both the schema type name (first, to maintain back compat), or the go type name if the former isn't found.
